requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
click==8.2.1
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
fastapi==0.115.12
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
Jinja2==3.1.6
jinja2-time==0.2.0
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydyf==0.11.0
pyparsing==3.2.3
pyphen==0.17.2
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
